backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "quiz-app-backend"
version = "0.5.1"
description = "QuizApp FastAPI backend — create, take, and track quizzes with AI generation"
requires-python = ">=3.12"
license = "MIT"
authors = [{ name = "QuizApp Team" }]
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: FastAPI",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
]

dependencies = [
    # Core Framework
    "fastapi==0.117.1",
    "uvicorn==0.37.0",
    "starlette==0.48.0",
    # Database
    "pymongo==4.10.1",
    # Data Validation & Settings
    "pydantic==2.11.9",
    "pydantic-settings==2.12.0",
    "pydantic_core==2.33.2",
    # Authentication & Security
    "python-jose[cryptography]==3.5.0",
    "passlib[bcrypt]==1.7.4",
    "bcrypt==4.2.1",
    "cryptography==43.0.3",
    # Rate Limiting
    "slowapi==0.1.9",
    "limits==5.8.0",
    # AI / Gemini
    "google-genai>=1.64.0",
    # Utilities
    "python-dotenv==1.2.1",
    "python-multipart==0.0.20",
    "anyio==4.12.1",
    "typing_extensions==4.12.2",
]

[dependency-groups]
dev = [
    "pytest==8.4.2",
    "pytest-asyncio==1.0.0",
    "httpx==0.28.1",
    "pytest-cov==5.0.0",
    "ruff==0.9.9",
    "pre-commit==4.2.0",
]

# ── Build ────────────────────────────────────────────────────────────────────
[tool.hatch.build.targets.wheel]
packages = ["app"]

# ── Ruff ─────────────────────────────────────────────────────────────────────
[tool.ruff]
target-version = "py312"
line-length = 100
src = ["app", "tests"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "UP",   # pyupgrade
    "S",    # flake8-bandit (security)
    "RUF",  # ruff-specific
]
ignore = [
    "S105",  # hardcoded password (false positives on defaults)
    "S106",  # hardcoded password argument
    "B008",  # function call in default argument (FastAPI Depends)
]

[tool.ruff.lint.isort]
known-first-party = ["app"]

# ── Pytest ───────────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
